Create Mechansism for passing replica init tids to voter data structures
now that thread creation is being handeld outside of the voter code, and in the main.c file we need a way to pass those thread ids to the voter.

currently storing them in global memory and allowing rust to call out to c to get a pointer to this memory. There is certinaly a beter mechanism (build the thread creation into rust) but this seems to work for now.

removed demo app rust file -> in the process of translating that to composite components
removed voter thd creation